// src/store/generalPropReducer/generalProp.types.ts
export enum GENERAL_PROPS_ACTION_TYPES {
  SET_THEME = "SET_THEME",
  SET_PATH = "SET_PATH",
  TOGGLE_USER_MENU = "TOGGLE_USER_MENU",
  TOGGLE_PRODUCT_CARD_MENU = "TOGGLE_PRODUCT_CARD_MENU",
  TOGGLE_SORTING_IS_IN_VIEW = "TOGGLE_SORTING_IS_IN_VIEW",
  SET_CURRENT_PRODUCT_CARD = "SET_CURRENT_PRODUCT_CARD",

  SET_VIEW_LIMITER = "SET_VIEW_LIMITER",
  SET_DELIVERY = "SET_DELIVERY",

}

export type PathType = string;